Add countdown and gated page
