+++ package.json
@@ -11,6 +11,7 @@
     "redux": "^4.1.1",
     "redux-thunk": "^2.3.0",
     "styled-components": "^5.3.0",
+    "styled-system": "^5.1.5",
     "web-vitals": "^1.0.1",
     "web3": "^1.5.1",
     "web3-eth-contract": "^1.5.2"
