+++ src/utils/leadZeros.js
@@ -1,0 +1,7 @@
+export default (num) => {
+  if (num < 10) {
+    return `0${num}`;
+  }
+
+  return num;
+};
